+++ examples/contracts/tokens/Colors.sol
