+++ examples/contracts/tokens/QuickCoin.sol
