--- examples/tokens/.env.example
@@ -1,3 +1,0 @@
-ETHERSCAN_API_KEY=<YOUR_ETHERSCAN_API>
-ROPSTEN_URL=https://eth-ropsten.alchemyapi.io/v2/<YOUR ALCHEMY KEY>
-MNEMONIC=<YOUR_MNEMONIC>
