--- examples/tokens/.gitignore
@@ -1,10 +1,0 @@
-node_modules
-.env
-coverage
-coverage.json
-typechain
-
-#Hardhat files
-cache
-artifacts
-deployments
