--- examples/tokens/.npmignore
@@ -1,3 +1,0 @@
-hardhat.config.ts
-scripts
-test
