--- examples/tokens/.prettierignore
@@ -1,5 +1,0 @@
-node_modules
-artifacts
-cache
-coverage*
-gasReporterOutput.json
