--- examples/tokens/deploy/001_deploy_quickcoin.ts
@@ -1,19 +1,0 @@
-import { HardhatRuntimeEnvironment } from "hardhat/types";
-import { DeployFunction } from "hardhat-deploy/types";
-
-const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
-  const { deployments, getNamedAccounts } = hre;
-  const { deploy } = deployments;
-
-  const { deployer } = await getNamedAccounts();
-
-  await deploy("QuickCoin", {
-    from: deployer,
-    args: [],
-    log: true,
-  });
-};
-
-export default func;
-
-func.tags = ["QuickCoin"];
