--- examples/tokens/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "hardhat-project",
-  "devDependencies": {
-    "@nomiclabs/hardhat-ethers": "npm:hardhat-deploy-ethers@^0.3.0-beta.13",
-    "@nomiclabs/hardhat-etherscan": "^2.1.8",
-    "@nomiclabs/hardhat-waffle": "^2.0.1",
-    "@typechain/ethers-v5": "^7.2.0",
-    "@typechain/hardhat": "^2.3.1",
-    "@types/chai": "^4.3.0",
-    "@types/mocha": "^9.0.0",
-    "@types/node": "^12.20.41",
-    "@typescript-eslint/eslint-plugin": "^5.10.0",
-    "@typescript-eslint/parser": "^5.10.0",
-    "chai": "^4.3.4",
-    "dotenv": "^10.0.0",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^3.4.1",
-    "eslint-plugin-promise": "^5.2.0",
-    "ethereum-waffle": "^3.4.0",
-    "ethers": "^5.5.3",
-    "hardhat": "^2.8.2",
-    "hardhat-deploy": "^0.9.26",
-    "prettier": "^2.5.1",
-    "prettier-plugin-solidity": "^1.0.0-beta.13",
-    "solhint": "^3.3.6",
-    "solidity-coverage": "^0.7.17",
-    "ts-node": "^10.4.0",
-    "typechain": "^5.2.0",
-    "typescript": "^4.5.4"
-  },
-  "scripts": {
-    "compile": "hardhat compile",
-    "test": "hardhat test --deploy-fixture",
-    "lint:check": "eslint '**/*.{js,ts}'",
-    "lint:fix": "eslint '**/*.{js,ts}' --fix",
-    "solhint:check": "npx solhint 'contracts/**/*.sol'",
-    "solhint:fix": "npx solhint 'contracts/**/*.sol' --fix",
-    "prettier:check": "prettier '**/*.{json,sol,md}' --check",
-    "prettier:fix": "prettier '**/*.{json,sol,md}' --write",
-    "deploy:ropsten": "hardhat --network ropsten deploy",
-    "verify:ropsten": "hardhat --network ropsten etherscan-verify"
-  },
-  "dependencies": {
-    "@openzeppelin/contracts": "^4.4.2"
-  }
-}
