--- examples/tokens/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2018",
-    "module": "commonjs",
-    "strict": true,
-    "esModuleInterop": true,
-    "outDir": "dist",
-    "declaration": true
-  },
-  "include": ["deploy", "./deploy", "./test", "./typechain"],
-  "files": ["./hardhat.config.ts"]
-}
